new depadency packages
Former-commit-id: ba4315f0c2035f991536677e660298d740de857c [formerly 931f67bde30ab7a04d93b44d1af3fcb8d92c85d4]
Former-commit-id: 54cd368ffb23375fd2e07249a67cb87e678e8a8d

+++ ExViewer/ViewModels/GalleryVM.cs
@@ -327,7 +327,7 @@ namespace ExViewer.ViewModels
                 var prop = await current.ImageFile.GetBasicPropertiesAsync();
                 var imageProp = await current.ImageFile.Properties.GetImagePropertiesAsync();
                 this.CurrentInfo = string.Format(Strings.Resources.Views.ImagePage.ImageFileInfo, current.ImageFile.DisplayType,
-                    Opportunity.Converters.ByteSizeToStringConverter.ByteSizeToString(prop.Size, Opportunity.Converters.UnitPrefix.Binary),
+                    Opportunity.Converters.ByteSizeToStringConverter.ByteSizeToString((long)prop.Size, Opportunity.Converters.UnitPrefix.Binary),
                     imageProp.Width.ToString(), imageProp.Height.ToString());
             });
         }
